pyproject.toml
[project]
name = "lazydatabricks"
version = "1.0.0"
description = "A keyboard-first TUI for Databricks ops — lazygit for your data platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Fahad Baig"}
]
keywords = ["databricks", "tui", "terminal", "devops", "data-engineering", "clusters", "jobs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: System :: Systems Administration",
    "Environment :: Console",
]

dependencies = [
    "databricks-sdk>=0.38.0",
    "textual>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
spark = [
    "databricks-connect>=16.1.0,<16.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[project.scripts]
lazydatabricks = "lazydatabricks.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lazydatabricks"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
